Added Navigation bar, get by location and google sign in

+++ app/searchdelete/searchdelete.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { UserserviceService } from '../userservice.service';
+import { NotificationService } from '../notification.service';
 
 @Component({
   selector: 'app-searchdelete',
@@ -7,11 +8,32 @@ import { UserserviceService } from '../userservice.service';
   styleUrls: ['./searchdelete.component.css']
 })
 export class SearchdeleteComponent implements OnInit {
+  loc:string;
 users:any;
-  constructor(private userService: UserserviceService) { }
+id:number;
+types=["id","location"];
+type: string;
+  constructor(private userService: UserserviceService,private notificationService: NotificationService) { }
 public deleteCandidate(id:number)
 {
-  let resp=this.userService.deleteById(id);
+  if(confirm('Are you sure to delete this the record?'))
+  {
+    this.notificationService.warn('! Deleted Successfully');
+    let resp=this.userService.deleteById(id);
+  resp.subscribe((data)=>this.users=data)
+  }
+  
+}
+public GetCandidateByLoc()
+{
+  let resp=this.userService.getUserByLocation(this.loc)
+  resp.subscribe((data)=>this.users=data)
+  console.log(this.users)
+}
+public GetCandidateById()
+{
+  console.log("Id getting called ")
+  let resp=this.userService.getUserById(this.id)
   resp.subscribe((data)=>this.users=data)
 }
   ngOnInit(): void {
